webapp: relocate Edit 3D header actions into the editor toolbar
The Edit 3D page stacked two top bars (the global header plus the 3D
editor's own toolbar), eating vertical space the viewport wants. On
desktop /edit-3d routes the global header is now hidden and its
right-side actions live inside the editor's canvas toolbar instead.

- Hide the global TopBar on desktop /edit-3d* (mobile keeps it for the
  gate's nav chrome).
- Extract the header's right cluster (pricing, credits, task queue,
  profile, and the credits/settings modals) into a reusable
  TopBarActions, rendered both in the global header and in the editor.
- Add optional topBarStartSlot / topBarEndSlot props to Stage3D and
  Stage3DBody; the webapp injects the scene title and TopBarActions.
  Tauri passes neither, so its toolbar is unchanged.
- Move the scene title under the File/Outliner/Shortcuts buttons and
  drop the dimmed "Edit Scene /" prefix; left-aligned, no shadow.
- Shift the demo "Rendered Output" overlay down so it clears the
  relocated top-right actions.
- Harden ButtonDropdown against a crash when the options array shrinks
  after an option index was selected.
- Align toolbar spacing (Controls3D, ControlsTopButtons) with the new
  layout.
